requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.1
fonttools==4.51.0
fpdf2==2.7.9
gitdb==4.0.11
GitPython==3.1.43
groq==0.6.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
md2pdf==1.0.1
mdurl==0.1.2
mutagen==1.47.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydyf==0.10.0
Pygments==2.18.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
tenacity==8.3.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
weasyprint==62.3
webencodings==0.5.1
websockets
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
zopfli==0.2.3
fastapi==0.95.1
pdfminer.six==20201018
